fix: split Favorite and FavoriteItem into two files
